+++ src/form/b-form/CHANGELOG.md
@@ -18,11 +18,13 @@ Changelog
 * Changed the root tag `div` to `form`
 * Removed the `form` ref and block
 * Changed the default value of `method` from `add` to `post`
+* Deprecated `ValidationError.el` and `ValidationError.validator`
 
 #### :rocket: New Feature
 
 * Added `getElValueToSubmit`
 * Added the `submitEnd` event
+* Added `ValidationError.component` and `ValidationError.error`
 
 #### :house: Internal